pyproject.toml
[tool.poetry]
name = "streamdeck-gui-ng"
version = "4.2.3"
description = "Next Generation Linux UI for the Elgato Stream Deck - Modernized fork with Python 3.11+ support"
authors = ["Timothy Crosley <timothy.crosley@gmail.com>"]
maintainers = ["millaguie <millaguie@users.noreply.github.com>"]
license = "MIT"
readme = "README.md"
packages = [
  {include = "streamdeck_ui"}
]


[tool.poetry.dependencies]
python = ">=3.11,<3.14"
streamdeck = "^0.9.5"
pillow = ">=10.3.0"
pyside6 = "^6.4.2"
CairoSVG = "^2.5.2"
filetype = "^1.0.10"
importlib-metadata = "^6.8.0"
evdev = "^1.6.1"
pyyaml = "^6.0"
requests = "^2.31.0"

[tool.poetry.group.docs.dependencies]
mkdocs = ">=1.4.2"
mkdocs-material = ">=9.1.14"

[tool.poetry.group.test.dependencies]
pytest = "^7.4.3"
pytest-cov = "^4.1.0"
pytest-mock = "^3.12.0"
pytest-qt = "^4.2.0"

[tool.poetry.group.lint.dependencies]
black = "^24.3.0"
flake8-bugbear = "^24.0.0"
mypy = "^1.6.1"
pep8-naming = "^0.13.3"
vulture = "^2.10"
bandit = "^1.7.5"
safety = "^3.7.0"
isort = "^5.12.0"


[tool.poetry.group.dev.dependencies]
types-pyyaml = "^6.0.12.20250915"

[tool.poetry.scripts]
streamdeck = "streamdeck_ui.gui:start"
streamdeckc = "streamdeck_ui.cli.server:execute"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120

[tool.isort]
profile = "hug"

[tool.mypy]
python_version = "3.11"

[tool.portray]
output_dir = "site"
include_reference_documentation = false
extra_dirs = ["art", "images", "media", "scripts"]

[tool.portray.mkdocs]
repo_url = "https://github.com/millaguie/streamdeck-gui-ng"
edit_uri = "edit/main/"
